requirements.txt
PyQt5
pandas
SpeechRecognition
google-generative-ai
python-dotenv
openpyxl
openai
pyttsx3